Deprecate FrozenDict/List, use tuples as node lists
These utility classes were a bit awkward and e.g. caused problems
with pickle since a list object is assumed to be appendable (#112).

Instead of FrozenLists, we now use homogeneous tuples to guarantee
that node lists and other collections are immutable. We also use the
Collection and Mapping types to indicate immutability.

Ordinary tuples are more lightweight and pythonic than FrozenLists.

We also exclude the links when pickling Tokens in order to make
pickling GraphQL schemas (which contain the tokens) less expensive.