When passing vars make sure `parse_literal` consumes at least 2 params
`variables` may be passed to a scalar type that does not do anything
with them while it is still a valid case we do not want to evaluate
to `Undefined`.

Example case: evaluation of a `GraphQLList` type that recurringly invokes `parse_literal`
passing the same `variables` regardless of whether the variables passed actually
are relevant for the scalar currently evaluated.